editoast: core_task: add `SimulationEnv`
* Adds the new `SimulationEnv` responsible for streaming simulation
results.
* Relies on `PathfindingEnv` to fetch paths to build the
simulation requests.
* Follows the same structure as `PathfindingEnv`,
but the `run` function is left for later.
* A few naming disparities
with `PathfindingEnv`. If they are accepted, `PathfindingEnv` naming
will be updated accordingly.

Two private submodules for organization:

`mod inputs`
-----

* Simulation parameters and consist, just like `PathfindingEnv`
* Electrical profiles are handled globally

Core requests for margins and power restrictions to be represented
differently, even though they're both essentially values between two
waypoints. The easiest way to represent that is by using a RangeMap,
making construction and iteration easy, coealescing adjacent ranges and
flattening overlapping ranges. All that simplifies both our code and
reduces the payload size. We also avoid relying on path item IDs that
only make sense in the RailJSON (present in the DB table as well), further
separating input schemas and high level representations of our objects.

Is also defined a `SimulationTrain` that ensures that pathfinding
and simulation parameters are consistent. It also helps with building
power restrictions and margins RangeMaps.

`mod request`
-----

Isolates the construction of the request requiring both the simulation
input described there above, and a path, provided by `PathfindingEnv`.

Most of the module is actually for testing the construction, even though
it's a fairly simple function thanks to the RangeMap representation.

Signed-off-by: Leo Valais <leo.valais97@gmail.com>