github: add cooldown for NPM dependabot upgrades
The NPM ecosystem is a bit of a trainwreck: regularly, some widely
used NPM packages are compromised. We absolutely don't need to
upgrade our dependencies as soon as a new version comes out,
instead wait for a bit to allow dust to settle and security issues
to be discovered.

See e.g. this article for a detailed motivation:
https://blog.yossarian.net/2025/11/21/We-should-all-be-using-dependency-cooldowns

Signed-off-by: Simon Ser <contact@emersion.fr>