core: try to distribute eng allowances in post-processing
In post-processing, we try to avoid any conflict by adding a
time point that matches the times during exploration. But
when the point is in an engineering allowance range, we don't
really know how much time to add.

This is a best guess that likely won't fix *every case*,
but it's better than only handling range boundaries.

Signed-off-by: Eloi Charpentier <eloi.charpentier.42@gmail.com>