core: drop Takes
The dependency upgrade has been broken for more than a year, and we
don't really use any of its functionality. What we do use has been
implemented as a really thin compatibility layer. Ideally, we'll remove
it too and fully embrace the fact that core is not an HTTP server
anymore.

Signed-off-by: Younes Khoudli <younes.khoudli@epita.fr>