front: adapt usePathStepsMetadata for the new op combobox
Since the new combobox will contain both the op name and the CH code, remove the logic to fetch all CH of the path step.
We display only the ch if one has been assigned (will be later displayed in the new combobox).
The new combobox will handle the fetch of the ch list with the search endpoint.

Signed-off-by: SharglutDev <p.filimon75@gmail.com>