editoast: remove ScopedBoxFuture and useless bound from transactions
ScopedFuture are useful if you have a need to keep two different
lifetime around for a future. Here, we were requiring the two lifetimes
to be identical... Which is pretty much a giveway that using a normal
Future would work just as well.

The core of the change is in database/src/db_connection_pool.rs, the
rest is just cleanup.

Signed-off-by: Younes Khoudli <younes.khoudli@epita.fr>