editoast, front: make paced field optional
editoast
- Adapt models and schemas
- Rework TrainId to better handle the merge of train schedule and paced
  trains

front
- Introduce temporary types of PacedTrains with paced field defined: PacedTrainWithPaced, PacedTrainResponseWithPaced
- Handle a paced train without paced field as a TrainSchedule (single train)
- Fix the conflicts/occupations endpoints and data workflow
- Fix labels when editing a pacedTrain without paced
- Fix RoundTrips suggestions for PacedTrains without paced
- Fix selected items labels to handle PacedTrains without paced
- Fix selectedTrainId process when creating or editing pacedTrains without paced field:
In useAutoSelectTrainId, if the train to select is a pacedTrain without paced, then the id to
select is its raw id, and not the id of its first occurrence. It's the same when editing a train.

The naming may not be ideal, but all this code will be cleaned in the coming weeks.

Signed-off-by: Florian Amsallem <florian.amsallem@gmail.com>
Co-authored-by: Clara Ni <clara.ni@outlook.fr>