front: wip: make curve layer generic
This is only the first step and it doesn't work yet. We need to remove the whole canvas concept, same as for the space-time chart, otherwise it won't work either.
The only CurveLayer components displayed here are speeds and eco-speeds: ETCS was removed here. Don't forget to call the component for the ETCS curves once the refacto is completed.

Signed-off-by: Erashin <alwenn.charpentier@gmail.com>