front: refacto getEditLabel
Since the PacedTrain without a paced field will soon replace the TrainSchedules,
we won't be able to distinguish single trains from paced trains only with
the id (it will always be PacedTrains).
We will then have to check the paced field of the timetableItem to do it.

To prepare this migration, getEditLabel now receives the timetableItem instead of simply
the timetableItemId.

Signed-off-by: Clara Ni <clara.ni@outlook.fr>