front: simplify logic about train updates
In the previous commits, the existing logic was kept:  when updating a
paced train without a paced field (i.e in trainSchedule mode), the existing
paced train was deleted and a new one was created (as if we were transforming
a paced train into a new train schedule, because the existing item has a PacedTrainId
and is considered as a paced train).

This workflow broke end-to-end tests.

Signed-off-by: Clara Ni <clara.ni@outlook.fr>