editoast: make the Error in the Result have only error variants
The `TrackOffsetResult` is an alias type of a complex `Result`.
The `Error` in this `Result` was `PathfindingResult` which is actually
an `enum` that can store some successful values. We actually only want
to return the error of this `enum` which is `Pathfinding::Failure`. So
this patch propose to change the API to use the sub-type directly.


Signed-off-by: Jean SIMARD <woshilapin@tuziwo.info>